Bug Fixes
There was an urgent issue with the Hourly Tick not firing due to a NPE
in the Trade Ship code.
Straight Forward, just updated the town.yml
Fix a bug with the Trade Ship Reporting the wrong money and taxes.
Fix an issue that caused extreme lag the more players that are online.
Fix an issue with Armor speed debuffs stacking.
Reduced the delay when a Wonder resumes building from up to 60 minutes
to up to 1 minute.