Bug Fixes and Performance Improvements
Move the player armor and road speed checks to be only on chunk borders
and on equip.
Fix an issue with Farms not growing after a while
Fix several issues with the Trade Ship.
Remove roads and walls from the undo task.
Fix the Mine level not adding bonus hammers
Force town and civ names to be ASCII printable (English base alpha)